Introduce `EmberCli::BuildMonitor`
The `BuildMonitor` class exposes two methods: `#check!` and `#reset`.

The `#check` method inspects the contents of the configured error file,
raising a `BuildError` if it isn't blank.

The `#reset` method will remove the error file when present.

This abstract moves the logic of monitoring for build errors out of
`EmberCli::App`, further simplifying the class and further minimizing
its responsibilities.
